fix chef deps
- make chef a development dependency of this gem

having the gem install try to upgrade chef and ohai causes quite a bit
of chaos.  strictly this gem does 'depend' on chef, but it makes no
sense to install it into a ruby which does not have chef and we really
don't want the side effect of having various shitty versions of rubygems
depsolver go and install chef and ohai.  its preserved as dev dep so
that bundle install will pick up chef for testing.

- bump the ruby version on travis to 2.0.0

chef-12 and ohai-8 no longer support ruby 1.9.3 so travis fails without
this.  it would be possible to use 1.9.3 and pin against chef 11 in
order to test against that, but its probably more important to start
testing against chef-12 on ruby >= 2.0.